Gracefully handle aggregates with missing metrics data
When a particular metric is not found for an aggregate query,
gnocchi currently just returns a 400 error. For a grafana graph
with multiple series this means grafana displays nothing at all.

This change makes gnocchi return an empty list of measures instead,
so grafana displays the other series as expected.

Change-Id: Ifc5fa7717cd07715fcda86682ac24d7384e09e2f